requirements.txt
# =============================================
# OpenReg – Python Dependencies
# =============================================

# Core
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.0
python-dateutil>=2.8.0
faker>=20.0.0
tqdm>=4.65.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Authentication & Security
bcrypt>=4.0.0
streamlit-authenticator>=0.4.0

# Database
psycopg2-binary>=2.9.0   # PostgreSQL driver
sqlalchemy>=2.0.0         # ORM for SQLite and PostgreSQL

# Monitoring & Logging
structlog>=23.0.0
prometheus-client>=0.17.0
python-json-logger>=2.0.0

# Performance & Caching
redis>=4.5.0
joblib>=1.3.0

# Testing  (was duplicated – kept once, unified version floor)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0